Rename: record폴더와 myPage폴더를 생성 -record 폴더 안에 record 관련 프래그먼트들 이동 -myPage 폴더 안에 myPage 관련 프래그먼트들 이동
