+++ app/src/main/java/com/jeong/sesac/sai/ui/myPage/SettingsFragment.kt
@@ -1,8 +1,6 @@
-package com.jeong.sesac.sai.ui
+package com.jeong.sesac.sai.ui.myPage
 
-import androidx.fragment.app.viewModels
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -10,7 +8,6 @@ import androidx.navigation.fragment.findNavController
 import com.jeong.sesac.sai.R
 import com.jeong.sesac.sai.databinding.FragmentSettingsBinding
 import com.jeong.sesac.sai.util.BaseFragment
-import com.jeong.sesac.sai.viewmodel.SettingsViewModel
 
 /** writer: 정지영
  *
