+++ app/src/main/java/com/jeong/sesac/sai/ui/record/MyRegisteredNotesFragment.kt
@@ -1,15 +1,11 @@
-package com.jeong.sesac.sai.ui
+package com.jeong.sesac.sai.ui.record
 
-import androidx.fragment.app.viewModels
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.jeong.sesac.sai.R
 import com.jeong.sesac.sai.databinding.FragmentMyRegisteredNotesBinding
 import com.jeong.sesac.sai.util.BaseFragment
-import com.jeong.sesac.sai.viewmodel.MyRegisteredNotesViewModel
 
 /** writer: 정지영
  *
