+++ app/src/main/res/layout/fragment_bookmarked_notes.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.BookmarkedNotesFragment">
+    tools:context=".ui.record.BookmarkedNotesFragment">
 
     <androidx.appcompat.widget.AppCompatTextView
         android:id="@+id/fragment_bookmarked_notes"
