+++ app/src/main/res/layout/fragment_completed_finds.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.CompletedFindsFragment">
+    tools:context=".ui.record.CompletedFindsFragment">
 
     <androidx.appcompat.widget.AppCompatTextView
         android:id="@+id/fragment_completed_finds"
