+++ app/src/main/res/layout/fragment_followers.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.FollowersFragment">
+    tools:context=".ui.myPage.FollowersFragment">
 
     <androidx.appcompat.widget.AppCompatTextView
         android:id="@+id/fragment_followers"
