+++ app/src/main/res/layout/fragment_login_management.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.LoginManagementFragment">
+    tools:context=".ui.myPage.LoginManagementFragment">
 
     <androidx.appcompat.widget.AppCompatTextView
         android:id="@+id/fragment_login_management"
