+++ app/src/main/res/layout/fragment_my_registered_notes.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.MyRegisteredNotesFragment">
+    tools:context=".ui.record.MyRegisteredNotesFragment">
 
     <androidx.appcompat.widget.AppCompatTextView
         android:id="@+id/fragment_my_registered_notes"
