+++ app/src/main/res/layout/fragment_settings.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.SettingsFragment">
+    tools:context=".ui.myPage.SettingsFragment">
 
     <androidx.appcompat.widget.AppCompatTextView
         android:id="@+id/fragment_settings"
